Move deprecated BaseModel into skimage tree.
