+++ src/skimage/measure/fit.py
@@ -1,5 +1,6 @@
+from warnings import warn
+
 from _skimage2.measure.fit import (
-    BaseModel as BaseModel,
     CircleModel as CircleModel,
     EllipseModel as EllipseModel,
     LineModelND as LineModelND,
@@ -7,6 +8,7 @@ from _skimage2.measure.fit import (
     add_from_estimate as add_from_estimate,
     ransac as ransac,
 )  # noqa: F401
+
 
 __all__ = [
     'BaseModel',
@@ -18,8 +20,22 @@ __all__ = [
     'ransac',
 ]
 
-from _skimage2.measure.fit import _dynamic_max_trials  # noqa: F401
+from _skimage2.measure.fit import (  # noqa: F401
+    _PARAMS_DEP_START,
+    _PARAMS_DEP_STOP,
+    _dynamic_max_trials,
+)
 
 from skimage._doctest_adapters import adapt_doctests
 
 adapt_doctests(globals())
+
+
+class BaseModel:
+    def __init_subclass__(self):
+        warn(
+            f'`BaseModel` deprecated since version {_PARAMS_DEP_START} and '
+            f'will be removed in version {_PARAMS_DEP_STOP}',
+            category=FutureWarning,
+            stacklevel=2,
+        )
